backend/requirements.txt
aniso8601==9.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
protobuf==3.15.6
pyrsistent==0.17.3
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
Werkzeug==0.16.1
